feat: insights-ansible-playbook-signer
* Card ID: CCT-691

Adds support for self-signed plays and revocation lists.

- Verifier now accepts hidden argument `--revocation-list`, which allows
  us to get around non-matching signatures of revoked plays that
  are embedded inside of the package.

Includes new setuptools entrypoint insights-ansible-playbook-signer.

The executable is not included in the RPM package.

The 'sanity' CI job creates a playbook using the Signer and then hacks
out the insights-client code to use custom GPG keys to verify it using
the legacy/reference implementation on CentOS Stream 9.